Split large files and flatten Public/ subdirectories
Split four oversized files into partial classes:
- CoreReplApp.Interactive.cs (1859) -> Interactive (628) + Autocomplete (1237)
- CoreReplApp.cs (1576) -> CoreReplApp (794) + Execution (806)
- ConsoleLineReader.cs (1425) -> ConsoleLineReader (598) + Autocomplete (833)
- HelpTextBuilder.cs (896) -> HelpTextBuilder (286) + Rendering (617)

Remove redundant Public/ subdirectories by moving their contents into
parent directories: Autocomplete, Interaction, Rendering, ShellCompletion,
and Terminal.
