+++ src/Repl.Core/Autocomplete/AutocompleteMode.cs
