+++ src/Repl.Core/Autocomplete/AutocompleteOptions.cs
