+++ src/Repl.Core/Interaction/AskMultiChoiceOptions.cs
