+++ src/Repl.Core/Interaction/AskSecretOptions.cs
