+++ src/Repl.Core/Interaction/IReplInteractionChannel.cs
