+++ src/Repl.Core/Interaction/IReplInteractionHandler.cs
