+++ src/Repl.Core/Interaction/IReplInteractionPresenter.cs
