+++ src/Repl.Core/Interaction/ITerminalInfo.cs
