+++ src/Repl.Core/Interaction/InteractionResult.cs
