+++ src/Repl.Core/Interaction/ReplClearScreenEvent.cs
