+++ src/Repl.Core/Interaction/ReplInteractionEvent.cs
