+++ src/Repl.Core/Rendering/AnsiMode.cs
