+++ src/Repl.Core/Rendering/AnsiPalette.cs
