+++ src/Repl.Core/Rendering/IAnsiPaletteProvider.cs
