+++ src/Repl.Core/ShellCompletion/ShellCompletionOptions.cs
