+++ src/Repl.Core/ShellCompletion/ShellCompletionSetupMode.cs
