+++ src/Repl.Core/ShellCompletion/ShellKind.cs
