+++ src/Repl.Core/Terminal/TerminalCapabilities.cs
